server/requirements.txt
python-dotenv
fastapi
uvicorn
pypdf
strands-agents[a2a]
strands-agents-tools[a2a_client]
litellm
cloudpathlib[s3,gs,azure]